bu/pyproject.toml
[project]
name = "bu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use[memory]>=0.1.45",
    "python-dotenv>=1.1.0",
]
